App now saves part of its current state (started/stopped) and adds all markers of the current DataSet when switching to the map

+++ sonification/app/src/main/java/mupro/hcm/sonification/database/SensorDataDao.java
@@ -37,5 +37,5 @@ public interface SensorDataDao {
     void delete(SensorData data);
 
     @Query("SELECT * FROM SensorData WHERE dataSetId=(:id)")
-    List<SensorData> getSensorDataForDataSet(final int id);
+    List<SensorData> getSensorDataForDataSet(final long id);
 }
